+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes will be tracked in this file
 ###Added
 - README
 - CHANGELOG
+- Intensity and seed values to keymap for NBrick
+
+###Changed
+- randVertsBMesh now explicitly uses range of -n to +n
+
+###Removed
+- randomizeVerts
 
 ##[Prehistoric] - 2017-01-24
 ###Added